циркулирующие алгоритмы для flc
