+++ handlers/flashcards/flashcards_managing.py
@@ -1,3 +1,5 @@
+import sqlite3
+
 from aiogram import types, Dispatcher
 from aiogram.dispatcher import FSMContext
 from aiogram.dispatcher.filters import Text
@@ -76,7 +78,7 @@ async def flashcards_managing_create_end(message: types.Message, state: FSMConte
     await message.answer(f'Показывать карточку с двух сторон? - {msg}')
     try:
         flashcard_dp_create(message.from_user.id, user_data["front"], user_data["back"], show_card)
-        await message.answer(f'Карточка успешно создана', reply_markup=types.ReplyKeyboardRemove())
+        await message.answer(f'Карточка успешно создана', reply_markup=flashcard_menu.get_keyboard_flashcard_start())
     except Exception:
         await message.answer(f'Что - то пошло не так, попробуйте снова')
     await state.finish()
@@ -89,7 +91,7 @@ async def flashcards_managing_del_start(message: types.Message):
     if len(all_cards) == 0:
         await message.answer(f'У вас нет карточек, которые вы могли бы удалалять',
                              reply_markup=types.ReplyKeyboardRemove())
-        await message.answer(f'Сначала создайте их', reply_markup=flashcard_menu.get_keyboard_flashcard_managing())
+        await message.answer(f'Сначала создайте их', reply_markup=flashcard_menu.get_keyboard_flashcard_start())
         return
 
     await message.answer(f'Чтобы удалить карточку - введите её id\n'
@@ -139,7 +141,7 @@ async def flashcards_managing_info(message: types.Message):
         await message.answer('У вас нет карточек')
         return
     else:
-        await message.answer('Все ваши карточки:', reply_markup=types.ReplyKeyboardRemove())
+        await message.answer('Все ваши карточки:', reply_markup=flashcard_menu.get_keyboard_flashcard_start())
         # Список всех карточек. Пример: [(54, "cat", "кошка"),(55, "dog", "собака")]
         all_cards = flashcard_dp_info(message.from_user.id)
         # Создание сообщения с информацией о всех каточках
@@ -148,6 +150,22 @@ async def flashcards_managing_info(message: types.Message):
             mes_print += f'{i + 1}:  {all_cards[i][1]}  -  {all_cards[i][2]}\n'
 
         await message.answer(mes_print)
+
+
+async def setting_show(message: types.Message):
+    CONN = sqlite3.connect('data_b/scipio.db')
+    cur = CONN.cursor()
+    msg = message.text
+    if msg == 'Показ карточек':
+        await message.answer('Вы можете настроить показ карточек(фото, текст)',
+                             reply_markup=flashcard_menu.setting_show())
+    elif msg == 'Фото':
+        cur.execute(f"""UPDATE users SET flc_show = 1 WHERE telegram_user_id = {message.from_user.id};""")
+        await message.answer("Показ карточке: Фото", reply_markup=flashcard_menu.get_keyboard_flashcard_start())
+    elif msg == 'Текст':
+        cur.execute(f"""UPDATE users SET flc_show = 0 WHERE telegram_user_id = {message.from_user.id};""")
+        await message.answer("Показ карточке: Текст", reply_markup=flashcard_menu.get_keyboard_flashcard_start())
+    cur.connection.commit()
 
 
 class FlashcardManaging(StatesGroup):
@@ -179,3 +197,4 @@ def register_handlers_flashcards_managing(dp: Dispatcher):
 
     dp.register_message_handler(flashcards_managing_del_end,
                                 state=FlashcardManaging.flashcards_managing_del_end)
+    dp.register_message_handler(setting_show, Text(["Показ карточек", "Фото", "Текст"]), state='*')
